+++ src/components/History.tsx
@@ -3,7 +3,8 @@ import type { FC } from 'react'
 import { css } from '@emotion/react'
 import type { Board, Position } from '@logic/board'
 import type { MoveTypes, Piece } from '@logic/pieces'
-import { useHistoryState } from '@pages/index'
+
+import { useHistoryState } from '@/state/history'
 
 export type History = {
   board: Board
