--- src/utils/messageHandler.tsx
@@ -1,7 +1,0 @@
-export default (io, socket) => {
-  const createdMessage = (msg) => {
-    socket.broadcast.emit(`newIncomingMessage`, msg)
-  }
-
-  socket.on(`createdMessage`, createdMessage)
-}
